fix for failing tests
